requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile pyproject.toml
#
aiohttp==3.8.1
    # via gql
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via aiohttp
botocore==1.27.21
    # via gql
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via
    #   aiohttp
    #   requests
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
gql[all]==3.3.0
    # via RunPodPy (pyproject.toml)
graphql-core==3.2.1
    # via gql
idna==3.3
    # via
    #   requests
    #   yarl
jmespath==1.0.1
    # via botocore
loguru==0.6.0
    # via RunPodPy (pyproject.toml)
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
munch==2.5.0
    # via RunPodPy (pyproject.toml)
python-dateutil==2.8.2
    # via botocore
requests==2.28.1
    # via
    #   gql
    #   requests-toolbelt
requests-toolbelt==0.9.1
    # via gql
ruamel-yaml==0.17.21
    # via RunPodPy (pyproject.toml)
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
six==1.16.0
    # via
    #   munch
    #   python-dateutil
tabulate==0.8.10
    # via RunPodPy (pyproject.toml)
tomli==2.0.1 ; python_version < "3.11"
    # via RunPodPy (pyproject.toml)
urllib3==1.26.9
    # via
    #   botocore
    #   gql
    #   requests
websockets==10.3
    # via gql
yarl==1.7.2
    # via
    #   aiohttp
    #   gql
